Calculator unit test

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -1,16 +1,16 @@
 package com.kodilla.basic_assertion;
 
 public class Calculator {
-    public int a;
-    public int b;
+    public double a;
+    public double b;
 
-    public int sum(int a, int b){
+    public double sum(double a, double b){
         return a + b;
     }
-    public int subtract(int a, int b){
+    public double subtract(double a, double b){
         return a - b;
     }
-    public int squared(int a){
+    public double squared(double a){
         return a * a;
     }
 }
